Add files via upload
Introduction 
This project entails developing an e-commerce website selling clothes that adhere to the MVC pattern. The site must include a home page, product page, login page, signup page and contact us page. Registration requires users to upload an image and requires passwords to be encrypted. Sessions and cookies should be closed/distorted upon logout. Logged-in users can edit profiles, change passwords or add products to their cart admins can add edit delete products and view the product page or order list. The product page includes a search bar allowing users to filter results by categories, prices, and ratings users may view the homepage without logging in but require credentials to add items to the cart. 

Aim 
This project aims to design and develop an E-commerce website that allows users to browse and purchase clothing items online, following the MVC pattern and best practices for web development. 

Objectives 

The following is the objective of this coursework: - 

To implement the MVC pattern in the design and development of the E-commerce website, with distinct packages for the Model, View, and Controller components. 

To implement a user-friendly interface that allows customers to easily navigate the website, search for clothes and make purchases.  

Utilize secure encryption and authentication methods for safeguarding customer data. 

To incorporate a robust database system capable of storing and easily retrieving information about items, users, orders and transactions. 

To ensure effective collaboration as a group, with a clear division of tasks and responsibilities, communication and resolution of any conflicts or issues.